Refactor evaluateContrast to use Rgba and ContrastResults types from results.ts
