+++ src/ui/evaluate-contrast.ts
@@ -2,29 +2,11 @@ import Color from 'colorjs.io';
 import { apcaToInterpolatedFont, apcaValidateFont } from "a11y-color-contrast";
 import { extractColors } from "extract-colors"
 import { decode } from "./decode-image";
-
-
-interface ForeRgba {
-  r: number;
-  g: number;
-  b: number;
-  a: number;
-}
-
-interface ContrastResults {
-  textColor: string;
-  backgroundColor: string;
-  contrastApca: number;
-  contrastWcag: number;
-  apcaInterpolatedFont: any;
-  apcaValidatedFont: Record<number, Record<number, boolean>>;
-  wcagPass: boolean;
-  apcaPass: boolean;
-}
+import { ContrastResults, Rgba } from '../types/results';
 
 const evaluateContrast = async (
   messageContent: Uint8Array,
-  foreRgba: ForeRgba,
+  foreRgba: Rgba,
   fontSize: number,
   nodeId: string
 ): Promise<ContrastResults> => {
@@ -47,12 +29,16 @@ const evaluateContrast = async (
   const contrastApca = backgroundColorOKHL.contrast(foregroundColorOKHL, "APCA");
   const contrastWcag = backgroundColorOKHL.contrast(foregroundColorOKHL, "WCAG21");
   const apcaInterpolatedFont = apcaToInterpolatedFont(contrastApca);
-  const apcaValidatedFont = apcaValidateFont(contrastApca, fontSize, 400);
+  const allowedFontSizes = [10,12,14,15,16,18,21,24,28,32,36,42,48,60,72,96] as const;
+  const validFontSize = allowedFontSizes.includes(fontSize as typeof allowedFontSizes[number])
+    ? fontSize as typeof allowedFontSizes[number]
+    : 16;
+  const apcaValidatedFont = apcaValidateFont(contrastApca, validFontSize, 400) as Record<number, Record<number, boolean>>;
 
   // Does this pass WCAG?
   const wcagPass = contrastWcag >= 4.5;
   // Does this pass APCA?
-  const apcaPass = apcaValidatedFont[fontSize][400] === true;
+  const apcaPass = apcaValidatedFont[fontSize]?.[400] === true;
 
   const contrastResults: ContrastResults = {
     textColor: foregroundColor.toString({ format: 'hex' }),
